pyproject.toml
[project]
name = "anca"
version = "0.1.0"
description = "Autonomous Niche Content Agent - AI-powered content generation system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "crewai",
    "crewai-tools",
    "langchain-community",
    "html2text",
    "robotexclusionrulesparser",
    "tenacity",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "langchain>=1.1.2",
    "langchain-text-splitters>=1.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "litellm>=1.80.8",
    "playwright>=1.56.0",
    "fastapi>=0.124.0",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.12.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "tools*"]
exclude = ["articles*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=agents",
    "--cov=tools",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
